Separate search classes into new modules
- improves abstraction of the system
